Island: Migrate AgentConfiguration resource into the new service
