+++ monkey/monkey_island/cc/resources/__init__.py
@@ -4,7 +4,7 @@ from .clear_simulation_data import ClearSimulationData
 from .island_log import IslandLog
 from .reset_agent_configuration import ResetAgentConfiguration
 from .propagation_credentials import PropagationCredentials
-from .agent_configuration import AgentConfiguration
+from monkey_island.cc.services.agent_configuration_service.flask_resources import AgentConfiguration
 from .agent_events import AgentEvents
 from .agent_plugins import AgentPlugins
 from .agent_plugins_manifest import AgentPluginsManifest
