+++ monkey/monkey_island/cc/services/agent_configuration_service/flask_resources/__init__.py
@@ -1,0 +1,1 @@
+from .agent_configuration import AgentConfiguration
