+++ monkey/monkey_island/cc/services/agent_configuration_service/flask_resources/agent_configuration.py
