+++ monkey/tests/unit_tests/monkey_island/cc/services/agent_configuration_service/flask_resources/test_agent_configuration.py
@@ -12,8 +12,10 @@ from tests.unit_tests.monkey_island.conftest import get_url_for_resource
 from common.agent_configuration import AgentConfiguration
 from common.types import JSONSerializable
 from monkey_island.cc.repositories import StorageError
-from monkey_island.cc.resources import AgentConfiguration as AgentConfigurationResource
 from monkey_island.cc.services import IAgentConfigurationService, PluginConfigurationValidationError
+from monkey_island.cc.services.agent_configuration_service.flask_resources import (
+    AgentConfiguration as AgentConfigurationResource,
+)
 
 AGENT_CONFIGURATION_URL = get_url_for_resource(AgentConfigurationResource)
